+++ src/main/java/https/github/com/wallas5h/LaskoMed/infrastructure/database/entity/ClinicEntity.java
@@ -18,7 +18,7 @@ public class ClinicEntity {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name="clinic_id")
-  private Integer clinicId;
+  private Long clinicId;
 
   @Column(name="name")
   private String name;
